+++ objc2/src/test_utils.rs
@@ -2,7 +2,7 @@ use core::ops::{Deref, DerefMut};
 use std::os::raw::c_char;
 use std::sync::Once;
 
-use crate::declare::{ClassDecl, ProtocolDecl};
+use crate::declare::{ClassBuilder, ProtocolBuilder};
 use crate::runtime::{Class, Object, Protocol, Sel};
 use crate::{ffi, Encode, Encoding, MessageReceiver};
 
@@ -75,11 +75,11 @@ pub(crate) fn custom_class() -> &'static Class {
         // The runtime will call this method, so it has to be implemented
         extern "C" fn custom_obj_class_initialize(_this: &Class, _cmd: Sel) {}
 
-        let mut decl = ClassDecl::root("CustomObject", custom_obj_class_initialize).unwrap();
+        let mut builder = ClassBuilder::root("CustomObject", custom_obj_class_initialize).unwrap();
         let proto = custom_protocol();
 
-        decl.add_protocol(proto);
-        decl.add_ivar::<u32>("_foo");
+        builder.add_protocol(proto);
+        builder.add_ivar::<u32>("_foo");
 
         extern "C" fn custom_obj_set_foo(this: &mut Object, _cmd: Sel, foo: u32) {
             unsafe {
@@ -121,22 +121,22 @@ pub(crate) fn custom_class() -> &'static Class {
 
         unsafe {
             let set_foo: extern "C" fn(&mut Object, Sel, u32) = custom_obj_set_foo;
-            decl.add_method(sel!(setFoo:), set_foo);
+            builder.add_method(sel!(setFoo:), set_foo);
             let get_foo: extern "C" fn(&Object, Sel) -> u32 = custom_obj_get_foo;
-            decl.add_method(sel!(foo), get_foo);
+            builder.add_method(sel!(foo), get_foo);
             let get_struct: extern "C" fn(&Object, Sel) -> CustomStruct = custom_obj_get_struct;
-            decl.add_method(sel!(customStruct), get_struct);
+            builder.add_method(sel!(customStruct), get_struct);
             let class_method: extern "C" fn(&Class, Sel) -> u32 = custom_obj_class_method;
-            decl.add_class_method(sel!(classFoo), class_method);
+            builder.add_class_method(sel!(classFoo), class_method);
 
             let protocol_instance_method: extern "C" fn(&mut Object, Sel, u32) = custom_obj_set_bar;
-            decl.add_method(sel!(setBar:), protocol_instance_method);
+            builder.add_method(sel!(setBar:), protocol_instance_method);
             let protocol_class_method: extern "C" fn(&Class, Sel, i32, i32) -> i32 =
                 custom_obj_add_number_to_number;
-            decl.add_class_method(sel!(addNumber:toNumber:), protocol_class_method);
+            builder.add_class_method(sel!(addNumber:toNumber:), protocol_class_method);
         }
 
-        decl.register();
+        builder.register();
     });
 
     class!(CustomObject)
@@ -146,13 +146,13 @@ pub(crate) fn custom_protocol() -> &'static Protocol {
     static REGISTER_CUSTOM_PROTOCOL: Once = Once::new();
 
     REGISTER_CUSTOM_PROTOCOL.call_once(|| {
-        let mut decl = ProtocolDecl::new("CustomProtocol").unwrap();
+        let mut builder = ProtocolBuilder::new("CustomProtocol").unwrap();
 
-        decl.add_method_description::<(i32,), ()>(sel!(setBar:), true);
-        decl.add_method_description::<(), *const c_char>(sel!(getName), false);
-        decl.add_class_method_description::<(i32, i32), i32>(sel!(addNumber:toNumber:), true);
+        builder.add_method_description::<(i32,), ()>(sel!(setBar:), true);
+        builder.add_method_description::<(), *const c_char>(sel!(getName), false);
+        builder.add_class_method_description::<(i32, i32), i32>(sel!(addNumber:toNumber:), true);
 
-        decl.register();
+        builder.register();
     });
 
     Protocol::get("CustomProtocol").unwrap()
@@ -163,12 +163,12 @@ pub(crate) fn custom_subprotocol() -> &'static Protocol {
 
     REGISTER_CUSTOM_SUBPROTOCOL.call_once(|| {
         let super_proto = custom_protocol();
-        let mut decl = ProtocolDecl::new("CustomSubProtocol").unwrap();
+        let mut builder = ProtocolBuilder::new("CustomSubProtocol").unwrap();
 
-        decl.add_protocol(super_proto);
-        decl.add_method_description::<(u32,), u32>(sel!(calculateFoo:), true);
+        builder.add_protocol(super_proto);
+        builder.add_method_description::<(u32,), u32>(sel!(calculateFoo:), true);
 
-        decl.register();
+        builder.register();
     });
 
     Protocol::get("CustomSubProtocol").unwrap()
@@ -183,7 +183,7 @@ pub(crate) fn custom_subclass() -> &'static Class {
 
     REGISTER_CUSTOM_SUBCLASS.call_once(|| {
         let superclass = custom_class();
-        let mut decl = ClassDecl::new("CustomSubclassObject", superclass).unwrap();
+        let mut builder = ClassBuilder::new("CustomSubclassObject", superclass).unwrap();
 
         extern "C" fn custom_subclass_get_foo(this: &Object, _cmd: Sel) -> u32 {
             let foo: u32 = unsafe { msg_send![super(this, custom_class()), foo] };
@@ -192,10 +192,10 @@ pub(crate) fn custom_subclass() -> &'static Class {
 
         unsafe {
             let get_foo: extern "C" fn(&Object, Sel) -> u32 = custom_subclass_get_foo;
-            decl.add_method(sel!(foo), get_foo);
+            builder.add_method(sel!(foo), get_foo);
         }
 
-        decl.register();
+        builder.register();
     });
 
     class!(CustomSubclassObject)